+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
 
+### v0.1.1
+- InputFile Component (https://github.com/tony-dinh/react-toolkit/pull/10)
+- Skeleton Component (https://github.com/tony-dinh/react-toolkit/pull/11)
+- Card Component (https://github.com/tony-dinh/react-toolkit/pull/12)
+
 ### v0.1.0
 - Initial Release